+++ Pelitili/src/main/kotlin/gameaccount/AccountBalanceResponse.kt
@@ -1,6 +1,6 @@
 package gameaccount
 
 sealed class AccountBalanceResponse {
-    data class Success(val balance: Int) : AccountBalanceResponse()
+    data class Success(val balance: Long) : AccountBalanceResponse()
     data class Error(val errorMessage: String) : AccountBalanceResponse()
 }
