+++ Pelitili/src/main/kotlin/gameaccount/GameAccount.kt
@@ -2,8 +2,8 @@ package gameaccount
 
 interface GameAccount {
 
-    fun chargePlayerAccount(gameEventId:String, playerId: String, amount:Int): AccountBalanceResponse
-    fun depositPlayerAccount(gameEventId:String, playerId: String, amount:Int): AccountBalanceResponse
+    fun chargePlayerAccount(gameEventId:String, playerId: String, amount:Long): AccountBalanceResponse
+    fun depositPlayerAccount(gameEventId:String, playerId: String, amount:Long): AccountBalanceResponse
 
     //for illustration purposes
     fun getPlayerAccounts():List<String>
